requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
init-data-py==0.2.6
slowapi==0.1.9
